2.0 library - tolower - Improve syntax

+++ play.it-2/src/10_common.sh
@@ -64,14 +64,10 @@ print_error() {
 # USAGE: tolower $dir[…]
 tolower() {
 	for dir in "$@"; do
-		if [ ! -d "$dir" ]; then
-			return 1
-		fi
-		find "$dir" -depth | while read file; do
+		[ -d "$dir" ] || return 1
+		find "$dir" -depth -mindepth 1 | while read file; do
 			newfile="${file%/*}/$(echo "${file##*/}" | tr [:upper:] [:lower:])"
-			if [ ! -e "$newfile" ] && [ "$file" != "$dir" ]; then
-				mv "$file" "$newfile"
-			fi
+			[ -e "$newfile" ] || mv "$file" "$newfile"
 		done
 	done
 }
